[app] Update example description and move to camera package
