+++ app/src/main/AndroidManifest.xml
@@ -512,7 +512,7 @@
                 android:value=".ExampleOverviewActivity" />
         </activity>
         <activity
-            android:name=".examples.CameraPredefinedAnimatorsActivity"
+            android:name=".examples.camera.CameraPredefinedAnimatorsActivity"
             android:description="@string/activity_camera_predefined_animators_description"
             android:exported="true"
             android:label="@string/activity_camera_predefined_animators_title">
@@ -524,7 +524,7 @@
                 android:value=".ExampleOverviewActivity" />
         </activity>
         <activity
-            android:name=".examples.LowLevelCameraAnimatorActivity"
+            android:name=".examples.camera.LowLevelCameraAnimatorActivity"
             android:description="@string/description_camera_low_level"
             android:exported="true"
             android:label="@string/activity_camera_low_level">
